feat: docs - initial
