+++ .github/workflows/main.yml
@@ -10,9 +10,11 @@ jobs:
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v2
+      - run: cd docs
+      - run: pnpm install
+      - run: pnpm build-storybook
+      - run: cd ..
       - run: yarn install
-      - run: yarn add -D react react-dom
-      - run: yarn build-storybook
       - run: |
           git remote set-url origin https://git:${GITHUB_TOKEN}@github.com/${GITHUB_REPOSITORY}.git
           yarn deploy -- -u "github-actions-bot <support+actions@github.com>"
