+++ .gitignore
@@ -4,5 +4,5 @@
 
 .idea
 dist
-storybook-static
 node_modules
+yarn.lock
