--- docs/.eslintrc
@@ -1,24 +1,0 @@
-{
-	"parser": "@typescript-eslint/parser",
-	"plugins": [
-		"@typescript-eslint",
-		"prettier",
-		"react"
-	],
-	"env": {
-		"browser": true,
-		"jest": true
-	},
-	"parserOptions": {
-		"ecmaVersion": "latest",
-		"sourceType": "module",
-		"ecmaFeatures": {
-			"jsx": true
-		}
-	},
-	"settings": {
-		"react": {
-			"version": "detect"
-		}
-	}
-}
