+++ docs/.storybook/preview-head.html
@@ -1,0 +1,3 @@
+<script>
+	window.global = window
+</script>
