+++ docs/src/components/@shared/blur/thumbnail.jpeg
