+++ docs/src/components/accordion/accordion.stories.tsx
@@ -1,0 +1,7 @@
+// import { Accordion } from './accordion'
+
+const Accordion = () => <div>aaa</div>
+export default {
+	title: 'Advanced/Accordion',
+	component: Accordion
+}
