+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "transitions-kit",
-	"version": "1.1.8",
+	"version": "1.1.9",
 	"repository": {
 		"type": "git",
 		"url": "git+https://github.com/denchiklut/transitions-kit"
@@ -17,39 +17,36 @@
 	"description": "A set of predefined transition components",
 	"author": "Karl: ollylut@gmail.com",
 	"scripts": {
-		"start": "NODE_ENV=development webpack serve --mode development",
+		"postinstall": "husky install",
 		"build": "rm -rf dist && NODE_ENV=production && rollup -c",
-		"storybook": "start-storybook -p 6006",
-		"build-storybook": "build-storybook",
-		"deploy": "gh-pages -d storybook-static"
+		"start": "NODE_ENV=development webpack serve --mode development",
+		"format": "prettier --write \"src/**/*.{ts,tsx}\"",
+		"lint": "eslint --fix --quiet --ext .ts,.tsx ./"
 	},
 	"dependencies": {
 		"react-transition-group": "^4.4.5"
 	},
 	"peerDependencies": {
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0"
+		"react": ">=16.6.0",
+		"react-dom": ">=16.6.0"
 	},
 	"devDependencies": {
 		"@rollup/plugin-commonjs": "^22.0.1",
 		"@rollup/plugin-node-resolve": "^13.3.0",
 		"@rollup/plugin-typescript": "^8.3.3",
-		"@storybook/addon-a11y": "^6.5.10",
-		"@storybook/addon-essentials": "^6.5.10",
-		"@storybook/addon-interactions": "^6.5.10",
-		"@storybook/addon-links": "^6.5.10",
-		"@storybook/builder-webpack5": "^6.5.10",
-		"@storybook/manager-webpack5": "^6.5.10",
-		"@storybook/react": "^6.5.10",
-		"@storybook/testing-library": "^0.0.13",
 		"@types/react": "^18.0.15",
 		"@types/react-transition-group": "^4.4.5",
-		"gh-pages": "^4.0.0",
+		"@typescript-eslint/eslint-plugin": "^5.36.1",
+		"@typescript-eslint/parser": "^5.36.1",
+		"eslint": "^8.23.0",
+		"eslint-plugin-prettier": "^4.2.1",
+		"eslint-plugin-react": "^7.31.4",
+		"husky": "^8.0.1",
+		"lint-staged": "^13.0.3",
 		"prettier": "^2.7.1",
 		"rollup": "^2.77.2",
 		"rollup-plugin-peer-deps-external": "^2.2.4",
 		"rollup-plugin-terser": "^7.0.2",
-		"storybook-dark-mode": "^1.1.0",
 		"tslib": "^2.4.0",
 		"typescript": "^4.7.4"
 	},
@@ -61,6 +58,8 @@
 		"react",
 		"transition",
 		"uikit",
-		"transition-kit"
+		"zoom",
+		"grow",
+		"transitions-kit"
 	]
 }
