--- public/index.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="ru">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="initial-scale=1, width=device-width" />
-    <title>TransitionsKit</title>
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
-</html>
