+++ src/index.ts
@@ -2,3 +2,5 @@ export * from './fade'
 export * from './blur'
 export * from './collapse'
 export * from './slide'
+export * from './zoom'
+export * from './grow'
